% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sri.R
\name{attach.resource}
\alias{attach.resource}
\title{Create an Object that Makes Use of a Shared Resource}
\usage{
attach.resource(obj, ...)
}
\arguments{
\item{obj}{a descriptor object describing the location and type of a 
shared resource.}

\item{...}{other information needed for object-specific attaching.}
}
\value{
An R object that makes use of a share resource.
# Create a concrete function.
setMethod(
 "attach.resource", 
 signature(x="NULL"),
 function (obj, ...){
   NULL
 }
)

# Describe it
attach.resource(NULL)
}
\description{
This generic function should be used to attach to a 
system-wide shared resource on a computer. It is used by other packages.
}
