\name{describe}
\alias{describe}
\title{Create a description of a shared resource that can be attached}
\description{This function creates a description of a shared resource and is used in conjunction with \code{attach.resource} to utilize shared resources from an R session.}
\usage{
describe(x)
}
\arguments{
  \item{x}{The object to create a description for}
}
\value{A user defined description of a shared resource
}
\author{
Michael J. Kane
}
\seealso{\link{attach.resource}}
